+++ reptar/writers/__init__.py
@@ -3,8 +3,9 @@ from .ase_db import write_ase_db
 from .schnetpack_db import write_schnetpack_db
 from .writing_utils import string_xyz_arrays
 from .xyz_gap import write_xyz_gap
+from .xyz import write_xyz
 
 __all__ = [
     'write_pdb', 'write_ase_db', 'write_schnetpack_db', 'string_xyz_arrays',
-    'write_xyz_gap'
+    'write_xyz_gap', 'write_xyz'
 ]