+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Writer for standard XYZ files with optional comments.
 - Writer for extended XYZ files used in the Gaussian approximation potentials
 (GAP) code.
 - Documentation for writers.